requirements.txt
flask==3.1.0
fpdf2==2.8.3
